+++ src/Engine/Audio/SteamAudioInterface.cpp
@@ -1,6 +1,6 @@
 #include "SteamAudioInterface.hpp"
 
-void SteamAudioInterface::Process(float* input, size_t samples,  IPLVector3 * sourcePosition, float sourceRadius) {
+void SteamAudioInterface::Process(float* input, uint32_t samples,  IPLVector3 * sourcePosition, float sourceRadius) {
 
     //Create a steam audio buffer
     IPLAudioFormat format;
@@ -19,7 +19,7 @@ void SteamAudioInterface::Process(float* input, size_t samples,  IPLVector3 * so
     sAudio.Process(buf, 0, 0, 0, 0, 0);
 }
 
-float * SteamAudioInterface::GetProcessed(size_t* numSamples) {
+float * SteamAudioInterface::GetProcessed(uint32_t* numSamples) {
     IPLAudioBuffer* finalBuf = new IPLAudioBuffer;
     sAudio.GetFinalMix(finalBuf, numSamples);
